92.24MHz: H2 Stack component
This commit is an experiment in making the two stacks in the H2 core, the
variable and return stacks, into a separate component. This has been done, and
it works, although it really slows down the H2 core according to the timing
report (there seems to be no problem running the design on the FGPA). Unless
this can be sped up, these changes will not be merged back into the mainline.

This also removes a few warnings which were annoying (although they never
signified an actual problem).